% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coala_wrappers.R
\name{coalaSimThenDetectAndPlot}
\alias{coalaSimThenDetectAndPlot}
\title{Simulate and plot trees using the coala package}
\usage{
coalaSimThenDetectAndPlot(
  pop1_sample_size,
  pop2_sample_size,
  rate = 0.5,
  divergence.time = 5,
  loci_number = 100,
  loci_length = 1000,
  ploidy = 1,
  Ne1 = NA,
  Ne2 = NA,
  Ne1.time.change,
  Ne2.time.change,
  seed = NA,
  epsilon = 1,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{pop1_sample_size}{The size of the sample of the local population (population 1)}

\item{pop2_sample_size}{The size of the sample of the non-local population (population 2)}

\item{rate}{Migration rate from population 2 to 1 (default = 0.5)}

\item{divergence.time}{Time of divergence of population 2 from population 1 (default = 5)}

\item{loci_number}{The number of loci to add - see `coala::coal_model` (default = 100)}

\item{loci_length}{The length of the loci to add - see `coala::coal_model` (default = 1000)}

\item{ploidy}{The number of chromosomes that will be simulated per individual - see `coala::coal_model` (default = 1)}

\item{Ne1}{For specifying the effective population size of population 1. At the moment it's implemented quite crudely: if (as default) it is NA then the effective population size is constant over time, if not, it will change abruptly to the proportion specified here at the time given by `Ne1.time.change`}

\item{Ne2}{as above but for population 2}

\item{Ne1.time.change}{as above}

\item{Ne2.time.change}{as above}

\item{seed}{Option to set the seed (default NA)}

\item{epsilon}{Precision parameter to pass to DetectImports::test1 (default 1)}

\item{alpha}{The threshold for statistical significance (default 0.05)}

\item{...}{Extra options to pass to the ape plot.phylo function}
}
\value{
Plot
}
\description{
Simulate and plot trees using the coala package
}
