% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{detectImportsFAST}
\alias{detectImportsFAST}
\title{Quick test for the presence of imports}
\usage{
detectImportsFAST(tree, constant = F, online = F, epsilon, adjust = "none")
}
\arguments{
\item{tree}{Tree}

\item{constant}{Whether to assume that the local population size is constant}

\item{online}{Whether to perform the online test (ignored if contant=T)}

\item{epsilon}{Smoothing precision parameter (ignored if constant=T)}

\item{adjust}{Method for adjusting p-values (default is none)}
}
\value{
Results of importation test
}
\description{
Quick test for the presence of imports
}
