% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{keyStats}
\alias{keyStats}
\title{Compute dates of nodes, number of lineages through time, terminal branch lengths and coalescent intersection}
\usage{
keyStats(phy)
}
\arguments{
\item{phy}{Tree}
}
\value{
Tree including a matrix with dates of nodes, number of lineages through time and coalescent intersection
}
\description{
Compute dates of nodes, number of lineages through time, terminal branch lengths and coalescent intersection
}
