% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotCoalInt}
\alias{plotCoalInt}
\title{Plot the coalescent intervals}
\usage{
plotCoalInt(tree, ...)
}
\arguments{
\item{tree}{Tree}

\item{...}{Additional parameters are passed on}
}
\value{
Plot
}
\description{
Plot the coalescent intervals
}
