% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotImports}
\alias{plotImports}
\title{Plot the tree and imports}
\usage{
plotImports(
  tree,
  imports = c(),
  show.axis = T,
  colorBase = "black",
  colorImports = "red",
  colorDescendants = "blue",
  ...
)
}
\arguments{
\item{tree}{Phylogenetic tree, object of class phylo}

\item{imports}{Vector containing the indexes of imported tips}

\item{show.axis}{Whether or not to show the time axis}

\item{colorBase}{Base branch color}

\item{colorImports}{Color for imports}

\item{colorDescendants}{Color for the local descendants of imports}

\item{...}{Additional parameters are passed on to the plot.phylo function}
}
\value{
Plot of the tree with colored imports
}
\description{
Plot the tree and imports
}
