% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coala_wrappers.R
\name{plotTreeWithTipClasses}
\alias{plotTreeWithTipClasses}
\title{Plot a tree and label - if any - the non-local samples and putative imports}
\usage{
plotTreeWithTipClasses(tree, pop1_sample_size, alpha = 0.05, ...)
}
\arguments{
\item{tree}{Tree, as simulated using the coala package}

\item{pop1_sample_size}{The size of the sample of the local population}

\item{alpha}{The threshold for statistical significance (default 0.05)}

\item{...}{Extra options to pass to the ape plot.phylo function}
}
\value{
Plot
}
\description{
Plot a tree and label - if any - the non-local samples and putative imports
}
