% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.resDetectImports}
\alias{print.resDetectImports}
\title{Print function for DetectImports results}
\usage{
\method{print}{resDetectImports}(x, ...)
}
\arguments{
\item{x}{output from DetectImports test methods}

\item{...}{Passed on}
}
\value{
Print out details of DetectImports results
}
\description{
Print function for DetectImports results
}
