% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simCoal}
\alias{simCoal}
\title{Simulation of coalescent dated phylogeny given a fluctuating Ne(t) function}
\usage{
simCoal(dates = 1990:2010, NeFun = function(x) {
    return(10)
}, NeMin)
}
\arguments{
\item{dates}{Sampling dates}

\item{NeFun}{Population size function Ne(t)}

\item{NeMin}{Minimum value of Ne(t)}
}
\value{
A simulated dated phylogeny
}
\description{
Simulation of coalescent dated phylogeny given a fluctuating Ne(t) function
}
