% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simImports}
\alias{simImports}
\title{Simulation with imports}
\usage{
simImports(
  localPopStart = 2020,
  importDates = 2020.5,
  samplingStartDate = 2020,
  samplingEndDate = 2021,
  samplingNumber = 1000,
  globalNeg = 1,
  NeFunImp,
  computeKeyStats = T
)
}
\arguments{
\item{localPopStart}{Date when the local population was seeded}

\item{importDates}{Vector of dates when imports occurred}

\item{samplingStartDate}{Date when sampling starts}

\item{samplingEndDate}{Date when sampling ends}

\item{samplingNumber}{Number of genomes sampled}

\item{globalNeg}{Value of Ne*g for the global population}

\item{NeFunImp}{Function determining the demographic trajectory of imports}

\item{computeKeyStats}{Whether or not to compute the key stats, default is TRUE}
}
\value{
A simulated dated phylogeny
}
\description{
Simulation with imports
}
