% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.resDetectImports}
\alias{summary.resDetectImports}
\title{Summary function for DetectImports objects}
\usage{
\method{summary}{resDetectImports}(object, ...)
}
\arguments{
\item{object}{output from DetectImports test methods}

\item{...}{Passed on}
}
\value{
Print out details of DetectImports results
}
\description{
Summary function for DetectImports objects
}
