% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test2}
\alias{test2}
\title{ESD test for the presence of imports}
\usage{
test2(tree, epsilon, alpha = 0.05, maxi = round(Ntip(tree)/10), showPlot = F)
}
\arguments{
\item{tree}{Tree}

\item{epsilon}{Smoothing precision parameter}

\item{alpha}{Threshold for p-value significance, default is 0.05}

\item{maxi}{Maximum number of outliers in dataset}

\item{showPlot}{Whether to show a plot of the test}
}
\value{
p-values for importation
}
\description{
ESD test for the presence of imports
}
