% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{detectImports}
\alias{detectImports}
\title{Bayesian test for the presence of imports}
\usage{
detectImports(
  tree,
  model = "gp",
  adjust = "none",
  verbose = T,
  diagnostics = F,
  nchains = 4,
  iter = 4000,
  seed = NULL
)
}
\arguments{
\item{tree}{Tree}

\item{model}{One of 'constant' 'gp' 'gpcontam'}

\item{adjust}{Method for adjusting p-values (default is none)}

\item{verbose}{Whether to produce verbose output}

\item{diagnostics}{Whether to produce diagnostic warnings}

\item{nchains}{Number of chains to run in parallel}

\item{iter}{Number of iterations to run, with first quarter discarded}

\item{seed}{Seed}
}
\value{
Results of importation test
}
\description{
Bayesian test for the presence of imports
}
