% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.resDetectImports}
\alias{plot.resDetectImports}
\title{Plotting method for DetectImports results}
\usage{
\method{plot}{resDetectImports}(x, type = "scatter", ...)
}
\arguments{
\item{x}{Output from DetectImports test methods}

\item{type}{Type of plot to produce. Can be 'scatter' (default) or 'tree'}

\item{...}{Additional parameters are passed on}
}
\value{
Plot of DetectImports results
}
\description{
Plotting method for DetectImports results
}
