% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotImports}
\alias{plotImports}
\title{Plot the tree and imports}
\usage{
plotImports(
  tree,
  imports,
  showTimeAxis = T,
  colorBase = "black",
  colorImports = "red",
  colorDescendants = "blue",
  ...
)
}
\arguments{
\item{tree}{Tree}

\item{imports}{Imports}

\item{showTimeAxis}{Whether or not to show the time axis}

\item{colorBase}{Base branch color}

\item{colorImports}{Color for imports}

\item{colorDescendants}{Color for the local descendants of imports}

\item{...}{Additional parameters are passed on}
}
\value{
Plot
}
\description{
Plot the tree and imports
}
