% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test0}
\alias{test0}
\title{Test for the presence of imports assuming that the local population has a constant, known population size}
\usage{
test0(tree, Ne = NULL, adjust = "fdr", showPlot = T)
}
\arguments{
\item{tree}{Tree}

\item{Ne}{Population size}

\item{adjust}{Method for adjusting p-values (default is fdr)}

\item{showPlot}{Whether to show a plot of the test}
}
\value{
p-values for importation
}
\description{
Test for the presence of imports assuming that the local population has a constant, known population size
}
