% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test0}
\alias{test0}
\title{Test for the presence of imports assuming that the local population has a constant, known population size}
\usage{
test0(tree, Ne = NULL)
}
\arguments{
\item{tree}{Tree}

\item{Ne}{Population size}
}
\value{
p-values for importation
}
\description{
Test for the presence of imports assuming that the local population has a constant, known population size
}
