% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test1}
\alias{test1}
\title{Semi-parametric test for the presence of imports}
\usage{
test1(tree, epsilon, adjust = "fdr", showPlot = T)
}
\arguments{
\item{tree}{Tree}

\item{epsilon}{Precision parameter}

\item{adjust}{Method for adjusting p-values (default is fdr)}

\item{showPlot}{Whether to show a plot of the test}
}
\value{
p-values for importation
}
\description{
Semi-parametric test for the presence of imports
}
