% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test1}
\alias{test1}
\title{Semi-parametric test for the presence of imports}
\usage{
test1(tree, epsilon)
}
\arguments{
\item{tree}{Tree}

\item{epsilon}{Precision parameter}
}
\value{
p-values for importation
}
\description{
Semi-parametric test for the presence of imports
}
