% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{test1}
\alias{test1}
\title{Semi-parametric test for the presence of imports}
\usage{
test1(tree, epsilon, adjust = "fdr", online = F)
}
\arguments{
\item{tree}{Tree}

\item{epsilon}{Smoothing precision parameter}

\item{adjust}{Method for adjusting p-values (default is fdr)}

\item{online}{Whether to perform the online test}
}
\value{
p-values for importation
}
\description{
Semi-parametric test for the presence of imports
}
