% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testBayes}
\alias{testBayes}
\title{Bayesian test}
\usage{
testBayes(tree, adjust = "fdr")
}
\arguments{
\item{tree}{Tree}

\item{adjust}{Method for adjusting p-values (default is fdr)}
}
\value{
p-values for importation
}
\description{
Bayesian test
}
